refactor: split action.ts
